Wp-plugin for brands
This is a custom plugin developed for Shaping Nutrition that adds a custom post type with the name "Brands" that builds and maintains relationships with the Products Custom Post Type automatically added by Woocommerce. This plugin requires the free or PRO version of ACF to be installed as well as WooCommerce.